+++ tests/testthat/test-abname.R
@@ -5,7 +5,7 @@ test_that("abname works", {
   expect_equal(abname(c("AMOX", "GENT")), c("Amoxicillin", "Gentamicin"))
   expect_equal(abname(c("AMOX+GENT")), "Amoxicillin + gentamicin")
   expect_equal(abname("AMOX", from = 'umcg'), "Amoxicillin")
-  expect_equal(abname("amox", from = 'molis', tolower = TRUE), "amoxicillin")
+  expect_equal(abname("amox", from = 'certe', tolower = TRUE), "amoxicillin")
   expect_equal(abname("J01CA04", from = 'atc'), "Amoxicillin")
   expect_equal(abname(c("amox", "J01CA04", "Trimox", "dispermox", "Amoxil")),
                       rep("Amoxicillin", 5))